+++ src/motorman/index.js
@@ -1,12 +1,11 @@
 
-var { Motorman, DelegatesManager, DelegationsManager, Utilities } = require('./motorman');
-var { utilities } = require('./motorman');
+var { Motorman } = require('./motorman');
+var { DelegatesManager } = require('./delegates-manager');
+var { DelegationsManager } = require('./delegations-manager');
 
 module.exports = {
     maps: require('./maps'),
     Motorman,
     DelegatesManager,
     DelegationsManager,
-    Utilities,
-    utilities,
 };
