+++ src/motorman/motorman.js
@@ -28,383 +28,17 @@
  *  * Publish as NPM Module
  */
 var { ROUTER_METHOD_MAP, CONTROLLER_METHOD_MAP } = require('./maps');
-const Dictionary = Map;
-
-var channels = {
-    // Delegations
-    ['ROUTES:ESTABLISHED']: 'motorman://established/routes',
-    ['POLICIES:ESTABLISHED']: 'motorman://established/policies',
-    ['DELEGATIONS:ESTABLISHED']: 'motorman://established/delegations',
-    ['DELEGATIONS:READY']: 'motorman://prepared/delegations',
-    ['DELEGATIONS:STABLE']: 'motorman://stabilized/delegations',
-    // Delegates
-    ['MIDDLEWARE:ESTABLISHED']: 'motorman://established/middleware',
-    ['CONTROLLERS:ESTABLISHED']: 'motorman://established/controllers',
-    ['DELEGATES:ESTABLISHED']: 'motorman://established/delegates',
-    ['DELEGATES:READY']: 'motorman://prepared/delegates',
-    ['DELEGATES:STABLE']: 'motorman://stabilized/delegates',
-    // Motorman
-    ['READY']: 'motorman://prepared',
-    ['STABLE']: 'motorman://stabilized',
-};
-
-class EventHub {
-    
-    constructor() {
-        this.channels = { };
-        return this;
-    }
-    
-    publish() {}
-    subscribe() {}
-    unsubscribe() {}
-    
-}
-
-var Deferred = new (function Deferred() {
-    var _this = this;  // private context
-    var _resolve = null, _reject = null;
-    
-    function exe(resolve, reject) {
-        this.resolve = resolve;
-        this.reject = reject;
-    }
-    
-    class Deferred {
-        
-        constructor() {
-            var promise = new Promise( exe.bind(this) );
-            this.promise = promise;
-            return this;
-        }
-        
-        resolve(data) {
-            _this.resolve(data);
-            return this.promise;
-        }
-        reject(data) {
-            _this.reject(data);
-            return this.promise;
-        }
-        
-    }
-    
-    // export precepts
-    this.resolve = _resolve;
-    this.reject = _reject
-    
-    return Deferred;
-})();
-
-class Delegate {
-    
-    constructor(options = { name: '', type: undefined, instance: undefined }) {
-        var { name, type, instance } = options;
-        
-        this.name = name;
-        this.type = type;
-        this.instance = instance;
-        
-        return this;
-    }
-    
-}
-
-
-class Utilities {  // this should exist in its own package
-    
-    constructor() {}
-    
-    keyOf(route) {  // TODO: move to exports so that tests can use it
-        var { uri, method } = route, key = `${method} ${uri}`;
-        return key;
-    }
-    
-}
-
-const utilities = new Utilities();
-const sandbox = {
-    Superclass: EventHub,
-    Delegate,
-    Deferred,
-    channels,
-    utils: utilities,
-};
-
-
-var DelegatesManager = new (function DelegatesManager($) {
-    var _this = this;  // private context
-    var {
-        Superclass,
-        Delegate,
-        Deferred,
-        channels,
-        utils,
-    } = $;
-    var keyOf = utils.keyOf.bind(utils);
-    var _middleware = { }
-      , _controllers = { }
-      , _instances = { }
-      ;
-    var _dMiddleware = new Deferred()
-      , _dControllers = new Deferred()
-      , _dInstances = new Deferred()
-      ;
-    var _pMiddleware = _dMiddleware.promise
-      , _pControllers = _dControllers.promise
-      , _pInstances = _dInstances.promise
-      ;
-    
-    class DelegatesManager extends Superclass {
-        
-        get middleware() { return Array.from( this.$middleware.values() ); }
-        set middleware(value) { this.$middleware.clear(); value.forEach( (m) => this.$middleware.set(m.name, m) ); }
-        
-        get controllers() { return Array.from( this.$controllers.values() ); }
-        set controllers(value) { this.$controllers.clear(); value.forEach( (c) => this.$controllers.set(c.name, c) ); }
-        
-        get instances() { return Array.from( this.$instances.values() ); }
-        set instances(value) { this.$instances.clear(); value.forEach( (d) => this.$instances.set(d.name, d) ); }
-        
-        constructor(sandbox) {
-            super();
-            var $middleware = new Dictionary()
-              , $controllers = new Dictionary()
-              , $instances = new Dictionary()
-              ;
-            var pReady = Promise.all([ _this.pMiddleware, _this.pControllers ]).then( ([ m, c ]) => ({ middleware: m, controllers: c }) )
-              , pStable = Promise.all([ _this.pReady, _this.pInstances ]).then( ([ ready, instances ]) => ({ instances }) )
-              ;
-            
-            this.channels = channels;  // TODO: use injection
-            this.sandbox = sandbox;
-            this.$middleware = $middleware;
-            this.$controllers = $controllers;
-            this.$instances = $instances;
-            this.ready = pReady;
-            this.stable = pStable;
-            
-            return this.init();
-        }
-        init() {
-            // TODO: add event-hooks after require('@motorman/utilities');
-            this.ready
-                .then( () => this.load() )
-                ;
-            this.stable
-                .then( () => {} )
-                ;
-            return this;
-        }
-        
-        ['set:middleware'](definitions) {
-            var { $middleware } = this, definitions = { ...definitions }, collection = [ ];
-            
-            for (let name in definitions) collection.push( new Delegate({ name, type: definitions[name], instance: undefined }) );
-            this.middleware = collection;
-            this.instances = collection;
-            _this.dMiddleware.resolve($middleware);
-            
-            return this;
-        }
-        ['set:controllers'](definitions) {
-            var { $controllers } = this, definitions = { ...definitions }, collection = [ ];
-            
-            for (let name in definitions) collection.push( new Delegate({ name, type: definitions[name], instance: undefined }) );
-            this.controllers = collection;
-            this.instances = collection;
-            _this.dControllers.resolve($controllers);
-            
-            return this;
-        }
-        
-        load() {
-            var { $middleware, $controllers, $instances } = this;
-            for (let [ name, delegate ] of $middleware) this.instantiate(delegate, name, $middleware);
-            for (let [ name, delegate ] of $controllers) this.instantiate(delegate, name, $controllers);
-            _this.dInstances.resolve($instances);
-        }
-        instantiate(delegate, name, $map) {
-            var { sandbox } = this, { name, type: Class, instance: existing } = delegate, instance = new Class(sandbox);
-            var delegate = new Delegate({ ...delegate, instance });  // update `Delegate {...}` as `Delegate { ..., instance: Class {...} }`
-            this.$instances.set(name, delegate);
-            $map.set(name, delegate);
-        }
-        
-    }
-    
-    // export precepts
-    this.middleware = _middleware;
-    this.controllers = _controllers;
-    this.instances = _instances;
-    this.dMiddleware = _dMiddleware;
-    this.dControllers = _dControllers;
-    this.dInstances = _dInstances;
-    this.pMiddleware = _pMiddleware;
-    this.pControllers = _pControllers;
-    this.pInstances = _pInstances;
-    
-    return DelegatesManager;
-})(sandbox);
-
-var DelegationsManager = new (function DelegationsManager($) {
-    var _this = this;  // private context
-    var {
-        Superclass,
-        Delegate,
-        Deferred,
-        channels,
-        utils,
-    } = $;
-    var keyOf = utils.keyOf.bind(utils);
-    var _routes = [ ]
-      , _policies = [ ]
-      , _instances = [ ]
-      ;
-      ;
-    var _dRoutes = new Deferred()
-      , _dPolicies = new Deferred()
-      , _dInstances = new Deferred()
-      ;
-    var _pRoutes = _dRoutes.promise
-      , _pPolicies = _dPolicies.promise
-      , _pInstances = _dInstances.promise
-      ;
-    
-    function getDelegateInstances($store, names) {
-        var names = names.reduce( (array, n) => array.concat(n), [] )  // ensure no multidimentionality (flatten)
-          , names = names.filter( (n) => !!n )  // ensure !!~names.indexOf(undefined)
-          , delegates = names.map( (name) => $store.get(name) )
-          , delegates = delegates.filter( (d) => !!d )  // ensure !!~delegates.indexOf(undefined)
-          , instances = delegates.map( (delegate) => delegate.instance )
-          ;
-        return instances;
-    }
-    
-    function getHandlers(instances) {
-        var handlers = instances.map( (m) => m['execute'].bind(m) );
-        return handlers;
-    }
-    
-    class DelegationsManager extends Superclass {
-        
-        get routes() { return Array.from( this.$routes.values() ); }
-        set routes(value) { this.$routes.clear(); value.forEach( (r) => this.$routes.set(keyOf(r), r) ); }
-        
-        get policies() { return Array.from( this.$policies.values() ); }
-        set policies(value) { this.$policies.clear(); value.forEach( (p) => this.$policies.set(keyOf(p), p) ); }
-        
-        get instances() { return Array.from( this.$instances.values() ); }
-        set instances(value) { this.$instances.clear(); value.forEach( (p) => this.$instances.set(keyOf(p), p) ); }
-        
-        constructor(delegates) {
-            super();
-            var $routes = new Dictionary()
-              , $policies = new Dictionary()
-              , $instances = new Dictionary()
-              ;
-            var pReady = Promise.all([ _this.pRoutes, _this.pPolicies, delegates.stable ]).then( ([ routes, policies ]) => ({ routes, policies }) )
-              , pStable = Promise.all([ _this.pReady, _this.pInstances ]).then( ([ ready, instances ]) => ({ instances }) )
-              ;
-            
-            this.channels = channels;  // TODO: use injection
-            this.delegates = delegates;
-            this.$routes = $routes;
-            this.$policies = $policies;
-            this.$instances = $instances;
-            this.ready = pReady;
-            this.stable = pStable;
-            
-            return this.init();
-        }
-        init() {
-            // TODO: add event-hooks after require('@motorman/utilities');
-            this.ready
-                .then( () => this.generate() )
-                ;
-            this.stable
-                .then( () => {} )
-                ;
-            return this;
-        }
-        
-        ['set:routes'](routes) {
-            this.routes = routes;
-            _this.dRoutes.resolve(routes);
-            return this;
-        }
-        ['set:policies'](policies) {
-            this.policies = policies;
-            _this.dPolicies.resolve(policies);
-            return this;
-        }
-        
-        generate() {
-            var { routes } = this;
-            var instances = routes.map( (r) => this.map(r) );
-            
-            this.instances = instances;
-            _this.dInstances.resolve(instances);
-            
-            return this;
-        }
-        map(route) {
-            var { $policies, delegates } = this;
-            var { uri, method, controller, action, policies } = route;
-            
-            // begin @ controller-action
-            var key = keyOf(route), policy = $policies.get(key);
-            var pipeline = [ controller ]
-              , ware = getDelegateInstances(delegates.$instances, pipeline)
-              , [ instance ] = ware
-              , endpoint = instance[action].bind(instance)  // last hit handler
-              ;  // end @ controller-action
-            
-            // begin @ [specific] route-matching middleware
-            var pipeline = [].concat(policy.policies, policies)
-              , infraware = getDelegateInstances(delegates.$instances, pipeline)  // last hit middleware
-              ;  // end @ specific middleware handlers
-            
-            // begin @ [generic] catch-all middleware
-            var allMethodsKey =     keyOf({ method: '*', uri }),    allURIsKey =    keyOf({ method, uri: '*' }),    allKey =    keyOf({ method: '*', uri: '*' })
-              , absentMethodsKey =  keyOf({ method: '', uri }),     absentURIsKey = keyOf({ method, uri: '' }),     absentKey = keyOf({ method: '', uri: '' })
-              , keys = [ allMethodsKey, allURIsKey, allKey, absentMethodsKey, absentURIsKey, absentKey ]
-              ;  // end @ keys for policies { method: x, uri: y } e.g: { method: '*', uri: '*' } (run on all)
-            var pipeline = keys.map( (key) => $policies.get(key) )  // e.g: ['SomePolicy', ['SomePolicy']. undefined]
-              , supraware = getDelegateInstances(delegates.$instances, pipeline)  // first hit middleware
-              ;  // end @ generic middleware handlers
-            
-            var middleware = getHandlers([ ...supraware, ...infraware ]), handlers = [ ...middleware ];  // aggregate supra (1st) & infra (2nd)
-            var delegation = { uri, method, controller, action, policies, handlers };
-            
-            handlers.push(endpoint);  // push controller method to end of handlers pipeline (aggregate final handler in pipeline)
-            handlers.forEach( (handler) => !!handler.init && handler.init({ route }) );
-            
-            return delegation;
-        }
-        
-    }
-    
-    // export precepts
-    this.routes = _routes;
-    this.policies = _policies;
-    this.instances = _instances;
-    this.dRoutes = _dRoutes;
-    this.dPolicies = _dPolicies;
-    this.dInstances = _dInstances;
-    this.pRoutes = _pRoutes;
-    this.pPolicies = _pPolicies;
-    this.pInstances = _pInstances;
-    
-    return DelegationsManager;
-})(sandbox);
+var { sandbox } = require('./sandbox');
+var { DelegatesManager } = require('./delegates-manager');
+var { DelegationsManager } = require('./delegations-manager');
 
 var Motorman = new (function Motorman($) {
     var _this = this;  // private context
     var {
         Superclass,
-        Delegate,
         Deferred,
+        Dictionary,
+        Delegate,
         channels,
         utils,
     } = $;
@@ -490,4 +124,4 @@ var Motorman = new (function Motorman($) {
     return Motorman;
 })(sandbox);
 
-module.exports = { Motorman, DelegatesManager, DelegationsManager, Utilities, utilities };
+module.exports = { Motorman };
